requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pandas==2.2.3
pymongo==4.10.1
dnspython==2.7.0
pydantic==2.10.3
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.24.1